Verify the configuration in WConfd every time it's uploaded
Since currently for most operations WConfd receives the whole
configuration from Python jobs, it makes sense to verify it in every
such call. This won't be necessary when WConfd will implement
ConfigWriter operations internally.

Also expose the verification function as a RPC call and use it when
explicitly verifying the configuration. Later this call will replace the
Python verification completely.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>