Don't wait for running jobs on cancel
A running job will ignore a cancel request anyway
("Job ... is no longer waiting in the queue"), so
there is no point in waiting the full timeout to
see if the job terminates. However, we still send
the signal, to allow for future implementations of
clean cancellation of already running jobs. Note
that the clean up necessary in such a case will
still take longer than the timeout, so we won't
have to change the no-waiting practise.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>