Fix disk size computation in ComputeDiskSizePerVG
This patch fixes the disk size requirements computation in the
'ComputeDiskSizePerVG' method.

Currently, this method does not sum up the size requirements of all the
disks of an instance that reside in the same volume group, as intended.
Instead, it returns the size requirement of only the last disk of the
instance due to a wrong key application to the 'vgs' dictionary. This
patch fixes this issue.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>