Introduce the 'ComputeDisksInfo' method
This patch extends the 'cmdlib/instance_storage.py' module with a new
method, named 'ComputeDisksInfo'. This method is used by the disk
template conversion mechanism to compute the disks' information that
will be used for generating the new instance's template. In addition, it
modifies the arguments of the 'ComputeDisks' method to be used from the
'ComputeDisksInfo'.

The computed disks of the instance will match the size, mode and name of
the current instance's disks. It also computes the volume group and the
access parameter for the templates that support them. In addition, for
conversions targeting an extstorage disk template, the mandatory
provider's name or any arbitrary user-provided parameters will also be
included in the result. The output of this method will be used by the
'GenerateDiskTemplate' method to generate the new template of the
instance.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>