Apply 'ComputeDisksInfo' to LVM-based conversions
This patch applies the 'ComputeDisksInfo' method to the currently
supported conversions between the LVM-based disk templates, i.e., plain,
drbd.

It is preferred having a central method that computes the list of disks
that contain the information for the new disk template of an instance,
instead of manually building this list every time it is needed. This
method will also be used from the upcoming disk template conversion
mechanism.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>