Adapt 'CreateDisks' for disk template conversions
Currently, the 'all_node_uuids' variable in the 'CreateDisks' method is
computed from all the disks of the instance. This approach will not be
compatible with the upcoming disk template conversion feature. A use
case is the conversions targeting the 'drbd' template. In that case the
'all_node_uuids' will be wrongly computed from the current instance's
disks that do not contain the secondary node's uuid of the new disks,
causing them to be only created in the primary node of the instance.
Removing the 'instance_disks' argument and using the 'disks' argument
instead, solves this issue without affecting existing functionality.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>