Add 'disk_template' parameter to 'CreateDisks'
Modify the 'CreateDisks' method to accept the disk template of an
instance as an extra parameter.

This method will be also used by the conversion mechanism to create the
new disks of the instance. Since the disk template of the instance will
not have been modified at the time we create the new disks and
consequently calling this function, we pass the new disk template of the
instance as an extra argument. In case we do not specify the disk
template parameter, the instance's disk template will be used.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>