Add two new parameters to 'RemoveDisks'
Extend the 'RemoveDisks' method to accept the disk template, and the
disks of the instance we want to remove, as extra parameters.

This method will be used by the conversion mechanism to remove the old
disks of an instance after we successfully converted to the new disk
template. Instead of directly calling the 'blockdev_remove' RPC every
time we want to remove a sub-set of the instance's disks, we modify this
method accordingly to save repeated lines of code.

Besides the 'disks' parameter that specifies the disks we want to
remove, we also add the 'disk_template' parameter. At the time we remove
the old block devices, the instance has already converted to the new
template; in that case the old disk template must be passed as an extra
argument.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>