Lift 'blockdev_remove' from instance operations
This patch replaces the 'blockdev_remove' RPC call with the
'RemoveDisks' method, from the instance related operations in the
'cmdlib/instance.py' module.

The 'RemoveDisks' method will also be used from the upcoming disk
template conversion mechanism to remove the original block devices of
the instance and also for rolling back any modification made in case the
conversion operation fails. We use a central method for removing the
instance's disks instead of repeatedly calling the 'blockdev_remove'
every time we want to remove a set of disks, to save repeated lines of
code.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>