Warn about nodes policy wise in a bad state
Add a warning on startup of hbal, if nodes are in a policy-wise
bad state. Hbal will never do any moves where the accepting
node afterwards is in a policy-violating state. A consequence of
this is that hbal will not improve the situation if a cluster has
too many policy-wise bad states (issue 694). So make the user aware
of this situation.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>