Make htools ignore soft errors if desired
Make htools support the --ignore-soft-errors options,
telling to not care about soft errors during balancing.
This might be necessary if the cluster is so heavily
overloaded on oversubscribable resources that otherwise
no movement would be possible. Note that the balancing
metric is still designed to minimize the amount of
soft errors.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>