DRBD parser: consume initial empty resource lines
When a DRBD resource is taken down, its minor number is freed and its
place in /proc/drbd is taken by an empty line. Currently the parser is
able to consume any empty lines appearing inside the resource stanza,
but cannot handle the first resource being empty.

We fix this by consuming all EOL characters appearing before the device
number. A test for this case should also be added.

This fixes issue #869.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>