Clarify setup for parallel job execution performance test
Explicitly state, that parallelism enhancing options are to be used
in parallel performance tests, if it is best practice to use them on
the respective versions. A prerequisite for those options is that
using them does not require, even under heavy load but assuming enough
capacity in the cluster, any jobs to be retried.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>

Cherry-picked from a305cb96e24303414ed4512932be8208fc5fdcca.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>