Arbitrary MultiMap from all exported Constructors
Change the way MultiMap is an instance of Arbitrary to
be a random term of the exported functions. In this way,
we really obtain a sample of all values of type MultiMap
that are definable outside the module. In particular, we
verify that all invariants of the implementation are preserved
over all exported functions. This renders the extensionality
test meaningful.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>