Exclude archived jobs from configuration backup
For a cluster that existed for a long time, the archive of
old jobs might be big. Additionally, this data is neither
necessary for running new jobs, nor is it touched in the
upgrade procedure. Hence it is safe to be excluded from
the backup created on Ganeti version upgrades, thus obtaining
a smaller backup.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>