Collate adjacent standard lock requests
Collate standard lock requests (i.e., non-opportunistic, not in
panic mode) of adjacent levels to a single request to WConfD.
Besides reducing the amount of inter-process communication, the
main advantage is that fewer locks are held while we wait for
other locks. This is possible, as the new locking mechanism
allows to get all locks at once, once they're available.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>