Fix curious psutil lint error
To expose certain attributes, the psutil module performs python magic
that results in an attribute being present in the module, but not in
the __all__ list. As pylint uses the __all__ list to determine whether
property accesses are errors, it is understandably confused about this
state of affairs.

To fix this, we just use the newer cpu_count attribute which is
supposed to replace NUM_CPUS anyway.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>