Allow coexistence of adding and acquiring locks
The constraint that locks may not be added and acquired at the same
time was in place before 2.12 only due to the fact that additional
complicated logic was necessary to do so.

With the new locking system, implementing this becomes much easier,
and quite useful at that, as it will allow us to get rid of BGL usage
for instance renames a bit later.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>

Cherry-picked from: 9261433f5385f2dee65a8de7ce6f1baa0221c547

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>