Adapt relative path expansion in PATH to new environment
The test framework we use sets PYTHONPATH for tests so that
the python module under test is used. Unfortunately, the paths
in PYTHONPATH are relative paths. So we need to expand them for
this test. Here we just update our fragile work around, but the
only way the test is called is, well, in testing, so we can build
it for our test environment.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>