Allow LUs to declare lock levels required to be sequential
In many cases, an LU only needs a set of locks so that they
could as well be obtained all at once. In some cases, however,
the set of locks need at a particular lock level can only be
reliably determined if the locks at earlier levels are owned
already. The newly added instance variable dont_collate_locks
allows LUs to declare for which levels this is the case.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>