Build the squeeze schroot with profiling enabled
Since we're installing cabal from scratch, we can take the advantage of
it and build everything with profiling enabled (including libraries that
are installed during the installation of cabal).

Also since we're using a newer version of cabal, we can let it install
packages in parallel, speeding up the schroot build.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>