Create SSH keys outside Ganeti
If we run a QA with the --no-ssh-init option, we have to
make sure that the nodes can connect to each other before
the QA starts running. For this purpose, the QA code
distributes the master node's SSH key to all other nodes
(similar to the way Ganeti used to do it) if the
'modify_ssh_setup' is set to False for Ganeti itself.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>