Abstract the '_CalculateFileStorageDir' method
Currently, the '_CalculateFileStorageDir' method is only used by the
'LUInstanceCreate' LU to calculate the final file storage directory of
the instance that is created. Since the disk template conversion
mechanism allows conversions targeting a file-based disk template, this
method will also be used by the 'LUInstanceSetParams' LU.

This patch moves the '_CalculateFileStorageDir' method to the
'cmdlib/instance_storage.py' module to make it available to all logical
units dealing with the storage of instances.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>