Implement the disk template conversion feature
This patch provides support for converting between all the available
disk templates in Ganeti. It extends the 'CheckArguments' and
'CheckPrereq' methods of the 'LUInstanceSetParams' LU to cover the
necessary checks for the template conversion interface.

This patch does not provide support for conversions targeting the
'blockdev' template which acts only as a source template, since these
volumes are adopted pre-existent block devices. Also, conversions from
or to the 'diskless' template are not supported. The rest template
conversions are supported and use the '_ConvertInstanceTemplate'
interface. The only conversions that do not use this interface are the
ones between 'plain' <-> 'drbd', which fall back on Ganeti's existing
mechanism.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>