Improve feedback in '_ConvertInstanceTemplate'
This patch improves the feedback messages that the conversions interface
returns to the caller.

For conversions targeting an ExtStorage template, the provider's name is
also printed along with the disk template of the instance ('ext'). For
conversions where the 'ext' template is used as a source template we do
not include the same information but we print only the disk template,
since the disks of an ExtStorage instance may use different providers.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>