QA: add support for the disk template conversion
This patch provides support for testing the disk template conversion
feature between all the available disk templates.

The number of the disk templates that will be tested will be specified
by the user, and so we introduce a new option in the 'qa-sample.json'
file named 'convert-disk-templates'. The conversion test will run once
and will cover all the user-provided templates that are enabled in the
cluster. Since the conversion goes through a pipe, it suffices to test
every template once for import and once for export. This means that a
list of N templates to test, translates to N distinct conversion
operations.

Signed-off-by: Dimitris Bliablias <bl.dimitris@gmail.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>