Remove obsolete assertion
This code is scheduled to be removed any way (for a single job,
there is no point in managing a job queue) and the assertion does
not hold true any more in the new design. The problem with the
assertion is that it does not kill the process, so the job will
never be cleaned up. Hence remove the assertion and do a best effort
to get the job in an error state.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>