Add non-cherry-pickable changes related to state "rb----"
Ganeti versions 2.11 and 2.12 have both had patches related to strange
Xen domU states which are still legitimate running states, but Ganeti
did not recognize them as such. The previous two cherry-picked patches
bring some of these changes, but the final patch needed (79cb7eceaca2)
cannot be cherry-picked because of a conflict with the other two
patches. The changes were made on different branches, and merged
together at a later time.

This patch manually adds the change, merging together the descriptions
of why certain states are allowed, and putting all of this into the
function docstring.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>