Mount rootfs by LXC container
THe LXC can mount a block device as its rootfs by specifying the device
path as the lxc.rootfs in the config file.
The mount of rootfs will be done in an isolated mount namespace, so
after this change, we can't see the instance rootfs as a part of the
host file system without changing the mount namespace.
This patch changes the value of the lxc.rootfs config parameter to the
instance rootfs device path and discards procedures for
mouting/unmounting the rootfs.

Signed-off-by: Yuto KAWAMURA(kawamuray) <kawamuray.dadada@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>