Extend failed_by_timeout flag to RunResult
When using the timeout option of RunCmd, in some cases that we want to
know if a command failed by timing out or due to other issues.
This commit extends RunResult to hold an additional variable
"failed_by_timeout" which is set to True if a command was killed by
the timeout.

Signed-off-by: Yuto KAWAMURA(kawamuray) <kawamuray.dadada@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>