Add helper methods for passing fds via SCM_RIGHTS
..through qemu monitor sockets using fdsend.

For the hmp the `get-fd` command is used while for the qmp the
`add-fd` command is used, where the fd is added to the first
available fdset. Currently the first one is used during NIC hot-add
while the second will be used during disk hot-add.

Add also RemoveFdset() method for cleaning up fdsets.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Reviewed-by: Helga Velroyen <helgav@google.com>