Use WriteFile when possible in ssh.py
This is an additional patch to the SSH patch series
which simplifies the handling of public SSH keys by using
the utility function WriteFile as often as possible.
As it is a mess to merge it back into the series,
I am sending this as an additional patch at the end of the
series.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>