On upgrades, check for upgrades to resume first
Make gnt-cluster upgrade refuse to upgrade if an upgrade to be resumed
is present. For the convenience of the user, consider an upgrade
command to the same target version as the upgrade to be resumed as an
equivalent way of saying that the pending upgrade is to be resumed.
Fixes issue 906.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>