Clarify section separation in text format
While the old description ("The file should contain text data,
line-based, with single empty lines separating sections. The
lines themselves are column-based, with the pipe symbol (|) acting
as separator.") is precise and unambiguous, we did have some
misunderstandings in the past. So add the format of the empty
section (which makes sense to describe a cluster without any
instances) as a clarifying example.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>