Verify clean DATA_DIR on cluster init
A new Ganeti cluster should not inherit state from a previously existing
cluster. Therefore, on cluster initialization, verify that no dangerous
left-overs are present in the DATA_DIR. We cannot, however, insist on
the DATA_DIR being completely empty, as it is good practice to set up
the rapi user file ahead of time. What we therefore do is to verify
that
- the DATA_DIR, if present, only contains directory entries, and
- the QUEUE_DIR, if present, is empty.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>