Fix size handling in ExtStorage
Commit 9504487 introduced the Snapshot() method in ExtStrorage,
where the snap_size was erroneously exported to the external
script's environment without making it a string. This patch fixes
that in the following manner: The size/snap_size/grow of a disk are
integers. Let _ExtStorageEnvironment() be responsible for exporting
them as string values before invoking the external scripts.

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>