Fix LUInstanceSetParams prereq check wrt cpu_mask
Once I tried to modify a hvparam of the LXC hypervisor, I got this error:

    ValueError: max() arg is an empty sequence

This is because the current default value of the cpu_mask hypervisor
parameter of the LXC hypervisor is the zero-length string, so the
cpu_list returned by the ParseMultiCpuMask is empty.
The zero-length string is a valid value for the cpu_mask hvparam, so
we need to skip this check if the cpu_list is empty.

Signed-off-by: Yuto KAWAMURA(kawamuray) <kawamuray.dadada@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>