Fix BaseHypervisor: CheckParameterSyntax,ValidateParameters
Currently both of these functions check if the value satisfies
bool(value) == False.
Using this approach, numerical 0 would not be allowed as a value of a
hvparam.
We know that unspecified hvparams have a value of either None or the
zero-length string, so these functions need to check if the value is
the one of these unspecified values.

Signed-off-by: Yuto KAWAMURA(kawamuray) <kawamuray.dadada@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>