Add warning against hvparam changes with live migrations
Changing certain hvparams while an instance is running leads to a
dangerous situation for the instance. When the instance undergoes live
migration without having undergone a reboot that would have applied the
parameters, it is setup with the modified hvparams, and loaded with the
state relying on the old hvparams. Depending on what is used and how,
this might lead to a crash of the instance.

While a proper fix should go into a stable version of Ganeti as this is
not a trivial change, the least we can do is warn users that hvparam
changes may be dangerous.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>