Make userspace access checks properly strict
When a cluster- or group-level disk template access mode change is
specified, Ganeti makes sure that this change would not lead to an
invalid hypervisor/access mode combination on an existing instance.
This check was extended to non-RBD disk templates in patch 6488e5bc,
but was overly broad, targeting disk templates for which the
parameter was not set. This patch fixes the check.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>