Query.Filter: Add evaluateFilterM
This is a more general version of evaluateFilter, which can be
used for general evaluations of the Ganeti filter language that
do not involve ConfigData, live data and other specialties.

To make this function general enough for expressing very specific
evaluations (such as expressing `evaluateFilter`, which has various
short-cuts, in terms of `evaluateFilterM`), a GADT is introduced
that captures the operations we can perform on the leaves of the
query language.

This lack of separation between the recursive structure and the leaf
structure is what made it hard to write generic functions on `Filter`.

However, we currently need this separation only for filter evaluation
functions and not in the type of `Filter` itself, so
Ganeti.Query.Language remains unchanged.

Signed-off-by: Niklas Hambuechen <niklash@google.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>