WConfd: run as <prefix>-masterd by default
Since WConfd is taking over the old masterd role, it is best to have it
run as the old masterd user. A lot of the files (notably config.data and
client.pem) are accessible by masterd-only.

This is really just to ease the transition from masterd to wconfd and
get things working again and should be re-evaluated in the long run.

This fixes issues #952 and #953.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>