metad: run as root to bind port 80
metad needs to bind port 80, so it should really startup as root. Note
that this should ideally be fixed, either by changing the port to an
unprivileged one, or by dropping root privileges after binding the port.
The latter solution is the preferred one, since using port 80 is
mandated in the design document.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>