Extend the cancel-job luxi request by a kill flag
This flag will instruct luxid to send send a SIGKILL
instead of a SIGTERM if the job to be canceled has
already started. To not have to change RAPI all in
one big change, we keep the wire protocol backwards
compatible even though this is not strictly necessary
for the luxi protocol.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>