Compute migration restrictions and receiving ability
Provide two functions getMigRestrictions and getRecvMigRestrictions.
The first computes the migration restrictions, given the the cluster
tags and the tags of a node. The second computes the set of restrictions
a node is able to receive, again, given the cluster tags and the node
tags. Migration is possible if the the migration restrictions of the
source node are a subset of the set of restrictions the target node
is able to receive. This is as described in the location awareness
design document.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>