Remove misleading cleanup
The rule of thumb in the Ganeti QA is that we do not attempt to clean
up after ourselves - we clean up in detail before a QA, and attempt to
retain as much state as possible for debugging / hunting purposes.
This patch removes an instance removal which was meant to serve as
cleanup, but ended up reporting false errors.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>