Make parallel performance rename test more serial
While the rename operation still acquired the BGL lock, the parallel
performance rename test was effectively serialized. With the removal
of the BGL, the second rename fails as it hits a check for the
existence of the renamed instance before the rename is done.
This patch explicitly serializes the two operations by making them run
on success rather then submitting them almost immediately.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>