LXC: Rethink LXC instance directory structure
Since all lxc commands support the --lxcpath option which is used to
specify a lxc container definitions directory, we should follow the
directory structure expected by lxc utilities.
Currently we have no issues with the existing directory structure, but
once we try to introduce migration by using the lxc-checkpoint command,
this will be a problem.
This patch also creates a hierarchy for the instances directory instead
of using the top-level of the hypervisor's local directory.

Signed-off-by: Yuto KAWAMURA(kawamuray) <kawamuray.dadada@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>