Improve export version description and use the constant
So far Ganeti has had only one version of the description produced when
an instance is exported. Due to this, the code does not deal with
forward or backward compatibility right now, and assumes that we cannot
import an instance that does not have the same export version at stated
in the constants.

This requirement of version equality is introduced more explicitly in
this patch, by both enhancing the documentation of the constant, and
explicitly exporting the instances with the version stated in the
constants.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>