Generate domain-name unique names in genEmptyCluster
It is a prerequisite that all nodes in a cluster have different names.
In genEmptyCluster, this is achieved by adding a unique suffix to each
name. However, this still allows to have both, 'n-3' and 'n-3.n-4' as
names in the same cluster---and those names are considered equal, as
they might both the legitimate short hands of 'n-3.n-4.example.com'.

To avoid these kinds of name collisions, make genEmptyCluster only
use the first part of the FQDN and suffix it.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>