Document --no-node-setup option in man page
This patch documents the '--no-node-setup' option of
'gnt-node add', which prevents Ganeti from touching and
distributing the SSH keys of the new node. It also prevents
any manipulation of the 'authorized_keys' files of Ganeti.

This option has been around for some time, but was so far
not documented in the man page.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>