Update message when adding a node
When adding a node, Ganeti displays a warning about the
upcoming SSH manipulations. With the new SSH key handling,
this message is no longer correct and is updated in this
patch.

Besides that, this patch removes the restriction that
only SSH keys are generated when a node is not readded.
This makes sense from a security point of view, because
if a node comes back from repair one might not be sure
no untrusted personell was able to access the machine.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>