ConfigReader: Simplify `initConfigReader` argument passing
We passed in a cfg_transform and an IORef, just to apply the
cfg_transform and write it to the IORef.

Instead, directly pass in a function that decides what to do
with the read config, so that `ConfigReader` now is solely
responsible for reading the config, not for writing it somewhere.

Signed-off-by: Niklas Hambuechen <niklash@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>