Respect --no-node-setup in LUNodeAdd
While gnt_node.py already correctly complied to the
'--no-node-setup' option, the corresponding LU did
use the cluster's init flag '--no-ssh-init'. While not
advisable to do so, this would result in an error if
the cluster is initialized without the '--no-ssh-init'
option, but the node is added with the '--no-node-setup'
option.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>