Expose results of asynchronous actions using a data type
Previously waiting for the result of a call to an asynchronous worker
was tied to triggering the action. This patch adds a data type that
allows waiting and retrieving the result any time later (or never).

In particular, this allows functions that modify the lock set or the
configuration to return the Async data type, letting the caller trigger
different workers in parallel and decide if for what results to wait and
when.

This separation also slightly cleans the internal implementation.

Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>