RunNodeCmdWithStdin: convert flags to named parameters
This patch converts the boolean parameters of the function
'RunNodeCmdWithStdin' into named parameters to increase
readability. This fixes issue 975.

Note that it also changes the occurrences of the parameters
'verbose' and 'debug' to either 'False' or an option, because
'verbose' and 'debug' should not be 'True' on production
clusters to not leak security information. Ideally, all
occurences should be switchable as options on the commandline,
but which itself is out of the scope of this patch.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>