Force developer mode if haskell tests are enabled
While it is useful for users to be able to build ganeti
in an environment where some of the functions used in
our code base are already deprecated, it is important
that developers do use warnings as errors. As the haskell
tests are not necessary to build an run Ganeti, enabling
them is a clear sign that you're acting as a developer.
So imply developer mode in this case.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>