Support forthcoming instances
Change the type of instance objects to be the sum of real and
forthcoming instances. Note that the wire representation is
backwards compatible, so we can keep the change local to the
Haskell code base.

Also, for the time being, the monitoring daemon will collect
data only for real instances.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>