Also in tests, open fake config before editing
As our tests access the actual config object, the test
setup needs to follow the same protocol as for a real
configuration. We need to open it, before modifying,
even in the test setup. Note that the test setup is outside
the code to be tested, so should not acquire any locks
(hence we open the mock config in shared mode even though
we intend to modify it).

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Niklas Hambuechen <niklash@google.com>