Always OutDate() the lu's config
For historic reasons, each job has two, potentially
different, configuration references: one of the processor
and one for the logical unit. When requesting locks,
make sure we also out date the configuration of the logical
unit. Note that out dating is cheap as it only sets a flag
in the configuration; the actual refetch only happens
on next read access.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>

Cherry-picked from 8303156e9d35317be0fbf68fe147b3148874408c
ignoring the OutDates related to the lock-level collation
that was only added in a later branch.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>