Always OutDate() the lu's config
For historic reasons, each job has two, potentially
different, configuration references: one of the processor
and one for the logical unit. When requesting locks,
make sure we also out date the configuration of the logical
unit. Note that out dating is cheap as it only sets a flag
in the configuration; the actual refetch only happens
on next read access.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>